+++ native/comrak_nif/src/lib.rs
@@ -27,7 +27,7 @@ rustler::init!(
 
 #[rustler::nif(schedule = "DirtyCpu")]
 fn to_html(md: &str) -> String {
-    let inkjet_adapter = InkjetAdapter::new("onedark");
+    let inkjet_adapter = InkjetAdapter::default();
     let mut plugins = ComrakPlugins::default();
     plugins.render.codefence_syntax_highlighter = Some(&inkjet_adapter);
     markdown_to_html_with_plugins(md, &Options::default(), &plugins)
@@ -42,7 +42,13 @@ fn to_html_with_options<'a>(env: Env<'a>, md: &str, options: ExOptions) -> NifRe
     };
     match options.features.syntax_highlight_theme {
         Some(theme) => {
-            let inkjet_adapter = InkjetAdapter::new(&theme);
+            let inkjet_adapter = InkjetAdapter::new(
+                &theme,
+                options
+                    .features
+                    .syntax_highlight_inline_style
+                    .unwrap_or(true),
+            );
             let mut plugins = ComrakPlugins::default();
             plugins.render.codefence_syntax_highlighter = Some(&inkjet_adapter);
             let unsafe_html = markdown_to_html_with_plugins(md, &comrak_options, &plugins);
@@ -73,6 +79,62 @@ fn tree_to_html_with_options(tree: ExNodeTree, options: ExOptions) -> String {
     // FIXME: validate tree[0] is a document
     let node = tree.first().unwrap();
     node.format_document(&comrak_options)
+}
+
+fn extension_options_from_ex_options(options: &ExOptions) -> ExtensionOptions {
+    let mut extension_options = ExtensionOptions::default();
+
+    extension_options.strikethrough = options.extension.strikethrough;
+    extension_options.tagfilter = options.extension.tagfilter;
+    extension_options.table = options.extension.table;
+    extension_options.autolink = options.extension.autolink;
+    extension_options.tasklist = options.extension.tasklist;
+    extension_options.superscript = options.extension.superscript;
+    extension_options
+        .header_ids
+        .clone_from(&options.extension.header_ids);
+    extension_options.footnotes = options.extension.footnotes;
+    extension_options.description_lists = options.extension.description_lists;
+    extension_options
+        .front_matter_delimiter
+        .clone_from(&options.extension.front_matter_delimiter);
+    extension_options.multiline_block_quotes = options.extension.multiline_block_quotes;
+    extension_options.math_dollars = options.extension.math_dollars;
+    extension_options.math_code = options.extension.math_code;
+    extension_options.shortcodes = options.extension.shortcodes;
+    extension_options.wikilinks_title_after_pipe = options.extension.wikilinks_title_after_pipe;
+    extension_options.wikilinks_title_before_pipe = options.extension.wikilinks_title_before_pipe;
+
+    extension_options
+}
+
+fn parse_options_from_ex_options(options: &ExOptions) -> ParseOptions {
+    let mut parse_options = ParseOptions::default();
+
+    parse_options.smart = options.parse.smart;
+    parse_options
+        .default_info_string
+        .clone_from(&options.parse.default_info_string);
+    parse_options.relaxed_tasklist_matching = options.parse.relaxed_tasklist_matching;
+    parse_options.relaxed_autolinks = options.parse.relaxed_autolinks;
+
+    parse_options
+}
+
+fn render_options_from_ex_options(options: &ExOptions) -> RenderOptions {
+    let mut render_options = RenderOptions::default();
+
+    render_options.hardbreaks = options.render.hardbreaks;
+    render_options.github_pre_lang = options.render.github_pre_lang;
+    render_options.full_info_string = options.render.full_info_string;
+    render_options.width = options.render.width;
+    render_options.unsafe_ = options.render.unsafe_;
+    render_options.escape = options.render.escape;
+    render_options.list_style = ListStyleType::from(options.render.list_style.clone());
+    render_options.sourcepos = options.render.sourcepos;
+    render_options.escaped_char_spans = options.render.escaped_char_spans;
+
+    render_options
 }
 
 fn render(env: Env, unsafe_html: String, sanitize: bool) -> NifResult<Term> {